Move handleTeamChange as remote action
