+++ app/actions/websocket/teams.ts
@@ -3,7 +3,8 @@
 
 import {DeviceEventEmitter} from 'react-native';
 
-import {handleTeamChange, localRemoveUserFromTeam} from '@actions/local/team';
+import {removeUserFromTeam} from '@actions/local/team';
+import {fetchAllTeams, handleTeamChange} from '@actions/remote/team';
 import {updateUsersNoLongerVisible} from '@actions/remote/user';
 import Events from '@constants/events';
 import DatabaseManager from '@database/manager';
@@ -27,7 +28,8 @@ export async function handleLeaveTeamEvent(serverUrl: string, msg: any) {
     }
 
     if (user.id === msg.data.user_id) {
-        localRemoveUserFromTeam(serverUrl, msg.data.team_id);
+        await removeUserFromTeam(serverUrl, msg.data.team_id);
+        fetchAllTeams(serverUrl);
 
         if (isGuest(user.roles)) {
             updateUsersNoLongerVisible(serverUrl);
