+++ app/components/team_sidebar/team_list/team_item/team_item.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import {View} from 'react-native';
 
-import {handleTeamChange} from '@actions/local/team';
+import {handleTeamChange} from '@actions/remote/team';
 import Badge from '@components/badge';
 import TouchableWithFeedback from '@components/touchable_with_feedback';
 import {useServerUrl} from '@context/server';
